dataengineering/challenge-3/tester/requirements.txt
numpy
pandas
cassandra-driver